update: Created a JS file to get the files data
